backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.0
pydantic==2.5.0
pyswisseph==2.10.3.2
geopy==2.4.0
timezonefinder==6.2.0
python-dateutil==2.8.2
httpx==0.25.1